requirements.txt
wheel
django
requests
django-bootstrap-form
django-cors-headers
django-oauth-toolkit
django-widget-tweaks
djangorestframework
fpdf